+++ app/client/Http/Response.php
@@ -11,7 +11,7 @@ use PhpSPA\Core\Helper\FileHandler;
  *
  * @package HTTP
  * @author Samuel Paschalson <samuelpaschalson@gmail.com>
- * @copyright 2025 Samuel Paschalson
+ * @copyright 2026 Samuel Paschalson
  * @see https://phpspa.tech/references/response/#response-api-examples
  */
 class Response
@@ -261,9 +261,9 @@ class Response
 
       if ($fileType === 'text/html') {
          $contents = Compressor::compress($contents);
-      } elseif ($fileType === 'text/css' || $fileType === 'text/javascript' || $fileType === 'application/javascript') {
+      } elseif ($fileType === 'text/css') {
          $level = Compressor::getLevel();
-         $contents = Compressor::compressWithLevel($contents, $level, $fileType === 'text/css' ? 'CSS' : 'JS');
+         $contents = Compressor::compressWithLevel($contents, $level, $fileType);
          $contents = Compressor::gzipCompress($contents, $fileType);
       }