+++ app/core/Auth/Authentication.php
@@ -10,7 +10,7 @@ namespace phpspa\Core\Auth;
  * and other authentication schemes within the PhpSPA framework.
  *
  * @author dconco <me@dconco.tech>
- * @copyright 2025 Dave Conco
+ * @copyright 2026 Dave Conco
  * @license MIT
  */
 trait Authentication