+++ app/core/Helper/CallableInspector.php
@@ -10,7 +10,7 @@ use ReflectionFunction;
  * Useful for determining callable types, extracting reflection information,
  * and facilitating dynamic invocation or introspection of functions, methods, or closures.
  * @author dconco <me@dconco.tech>
- * @copyright 2025 Dave Conco
+ * @copyright 2026 Dave Conco
  * @static
  */
 class CallableInspector