+++ app/core/Helper/SessionHandler.php
@@ -12,7 +12,7 @@ use PhpSPA\Http\Session;
  * within PHP sessions in the PhpSPA framework.
  *
  * @author dconco <me@dconco.tech>
- * @copyright 2025 Dave Conco
+ * @copyright 2026 Dave Conco
  * @license MIT
  */
 class SessionHandler