+++ app/exceptions/AppException.php
@@ -11,7 +11,7 @@ namespace PhpSPA\Exceptions;
  *
  * @package PhpSPA\Exceptions
  * @author dconco <me@dconco.tech>
- * @copyright 2025 Dave Conco
+ * @copyright 2026 Dave Conco
  * @license MIT
  * @since v1.0.0
  */