+++ vite-template/index.php
@@ -1,8 +1,13 @@
 <?php
 
-require_once __DIR__ . '/../vendor/autoload.php';
+require_once __DIR__ . '/vendor/autoload.php';
 
 use PhpSPA\App;
+use PhpSPA\Compression\Compressor;
+use PhpSPA\Core\Http\HttpRequest;
+use PhpSPA\Http\Response;
+
+chdir(__DIR__); // --- Change the current working directory to the project root dir ---
 
 // --- Load components ---
 require_once 'app/layout/layout.php';
@@ -23,12 +28,32 @@ $app->attach($homePage);
 $app->attach($aboutPage);
 $app->attach($docsPage);
 
+// --- Add global meta data to the application ---
+
 $app->meta(charset: 'UTF-8')
     ->meta(name: 'viewport', content: 'width=device-width, initial-scale=1.0')
     ->link(rel: 'preconnect', content: 'https://fonts.googleapis.com')
-    ->link(rel: 'preconnect', content: 'https://fonts.gstatic.com', attributes: ['crossorigin' => ''])
-    ->link(rel: 'stylesheet', content: 'https://fonts.googleapis.com/css2?family=Space+Grotesk:wght@400;500;600;700&display=swap');
+    ->link(rel: 'preconnect', content: 'https://fonts.gstatic.com', attributes: ['crossorigin' => true])
+    ->link(rel: 'stylesheet', content: 'https://fonts.googleapis.com/css2?family=Space+Grotesk:wght@400;500;600;700&display=swap', type: 'text/css')
 
+    ->link(rel: 'shortcut icon', content: '/logo.svg', type: 'image/xml+svg')
+    ->link(rel: 'apple-touch-icon', content: '/logo.svg', type: 'image/xml+svg');
+
+
+// --- For Production ---
+
+if (getenv('APP_ENV') === 'production') {
+    $app->compression(Compressor::LEVEL_EXTREME);
+
+    // --- Apply Canonical Link ---
+    $req = new HttpRequest();
+    $siteURL = getenv('APP_URL') ?: $req->siteURL();
+
+    $app->link(rel: 'canonical', content: $siteURL);
+}
 
 // --- Run the application ---
 $app->run();
+
+// --- Return 404 error if no route was matched ---
+Response::sendError('Page not found', Response::StatusNotFound);